% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.alpha.R
\name{add.alpha}
\alias{add.alpha}
\title{Internal function for adding alpha values to plots}
\usage{
add.alpha(col, alpha = 1)
}
\arguments{
\item{alpha}{add.alpha()}
}
\description{
Internal function for adding alpha values to plots
}
