% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.ttest.R
\name{all.ttest}
\alias{all.ttest}
\title{All T.test Function}
\usage{
\method{all}{ttest}(sort = NULL, test = NULL, tresh = NULL,
  measurements = NULL, sessiontemp = NULL, stdout = TRUE,
  alphalevel = 0.1, absolutevalue = TRUE, standard = NULL,
  template = "Standard")
}
\arguments{
\item{sort}{sort}
}
\description{
All T.test Function
}
\examples{
all.ttest()
}
\keyword{all.ttest}

