% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_temp_space.R
\name{analytical_temp_space}
\alias{analytical_temp_space}
\title{Internal function to deal with temporary analytical directory and session temp}
\usage{
analytical_temp_space(output_options, sessiontempdir)
}
\arguments{
\item{sessiontempdir}{}
}
\description{
Internal function to deal with temporary analytical directory and session temp
}
\examples{
analytical_temp_space()
}
