% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antemortem_data.R
\docType{data}
\name{antemortem_data}
\alias{antemortem_data}
\title{This example data is in the format that .CSV templates should be when importing data into OsteoSort}
\description{
This example data is in the format that .CSV templates should be when importing data into OsteoSort
}
\keyword{data}
