% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antestat.input.R
\name{antestat.input}
\alias{antestat.input}
\title{antestat.input Input Function}
\usage{
antestat.input(antemortem_stature = NULL, postmortem_measurement = NULL,
  bone = "Femur", population = "Trotter-any-male", metric = "cm")
}
\arguments{
\item{antemortem_stature}{The antemortem stature for comparison}

\item{postmortem_measurement}{The postmortem measurement for comparison}

\item{bone}{The bone type to compare}

\item{population}{The population data to use for modeling}

\item{metric}{Specifies millimeters (mm), centimeters (cm), or inches (in) for stature}
}
\description{
Function to generate combinations of antemortem stature to postmortem measurements for use with antestat.regtest.R
}
\examples{
antestat.input()
}
\keyword{antestat}
