% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antestat.R
\name{antestat}
\alias{antestat}
\title{antestat Input Function}
\usage{
antestat(antemortem_stature = NULL, postmortem_measurement = NULL,
  bone = "Femur", prediction_level = 0.95, data = "Trotter_any_male",
  tails = 2)
}
\arguments{
\item{antemortem_stature}{The antemortem stature for comparison}

\item{postmortem_measurement}{The postmortem measurement for comparison}

\item{bone}{The bone type to compare}

\item{prediction_level}{The prediction interval level}

\item{data}{The population data to use for modeling}

\item{tails}{The number of tails for the t-statistic}
}
\description{
Function to produce a p-value evaluating the strength of evidence for comparing an antemortem stature to postmortem measurement
}
\examples{
antestat()
}
\keyword{antestat}
