% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antestat.regtest.R
\name{antestat.regtest}
\alias{antestat.regtest}
\title{antestat.regtest Input Function}
\usage{
antestat.regtest(sort = NULL, ref = NULL, sessiontempdir = NULL,
  output_options = c(FALSE, FALSE), prediction_interval = 0.95, tails = 2,
  alphalevel = 0.05, alphatest = TRUE, cores = 1)
}
\arguments{
\item{sessiontempdir}{Specifies temporary directory for analytical session}

\item{output_options}{C(TRUE,FALSE) First logic specifies excel output, second specifies plot output}

\item{prediction_interval}{The prediction interval level}

\item{tails}{The number of tails for the t-statistic}

\item{antemortem_stature}{The antemortem stature for comparison}

\item{postmortem_measurement}{The postmortem measurement for comparison}
}
\description{
Function to produce a p-value evaluating the strength of evidence for comparing an antemortem stature to postmortem measurement
}
\examples{
antestat.regtest()
}
\keyword{antestat}
