% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art.input.R
\name{art.input}
\alias{art.input}
\title{Articulation comparison Input Function}
\usage{
art.input(bone = NULL, sort = NULL)
}
\arguments{
\item{bone}{Specifies the bone type}

\item{sort}{Data to be sorted}
}
\description{
Articulation comparison Input Function
}
\examples{
art.input()
}
\keyword{art.input}
