% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art.ttest.R
\name{art.ttest}
\alias{art.ttest}
\title{Articulation comparison t-test Function}
\usage{
art.ttest(ref = NULL, sort = NULL, sessiontempdir = NULL, stdout = TRUE,
  cores = 1, alphalevel = 0.1, absolutevalue = TRUE,
  testagainstzero = FALSE, output_options = c(TRUE, FALSE), power = TRUE,
  plot = FALSE)
}
\arguments{
\item{ref}{Reference data}

\item{sort}{Sorted data for comparison}

\item{sessiontempdir}{Specifies temporary directory for analytical session if stdout is false}

\item{stdout}{If true, output will be data.frames only}

\item{cores}{Number of cores for parallel processing}

\item{alphalevel}{Specifies alpha level}

\item{absolutevalue}{Uses absolute value for D-values if true}

\item{testagainstzero}{Uses 0 for mean if true}

\item{output_options}{Uses two true and/or false values to specify output file types. c(TRUE,) uses .txt file per specimen and C(,TRUE) uses .csv output files}

\item{power}{If true, uses half-normal distribution power transformation}

\item{plot}{Used internally for OsteoShiny, do not call directly!}
}
\description{
Articulation comparison t-test Function
}
\examples{
art.ttest()
}
\keyword{art.ttest}
