% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_dist.R
\name{e_dist}
\alias{e_dist}
\title{Euclidean distance calculation between two matrices}
\usage{
e_dist(first_configuration, second_configuration)
}
\arguments{
\item{first_configuration}{The first pointcloud}

\item{second_configuration}{The second pointcloud

Modified from the distmat() function in the pracma package for R}
}
\description{
Euclidean distance calculation between two matrices
}
\keyword{e_dist}
