% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa.R
\name{efa}
\alias{efa}
\title{2D and 3D EFA function}
\usage{
efa(pointcloud, harmonics = 40)
}
\arguments{
\item{pointcloud}{The two-dimensional or three-dimensional point cloud}

\item{harmonics}{The number of harmonics

Modified from the efourier function in the Mmocs package for R and Hayley Fancourt's 3D EFA code from Honours thesis}
}
\description{
2D and 3D EFA function
}
\keyword{efa}
