% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment_margins.R
\name{fragment_margins}
\alias{fragment_margins}
\title{fragment_margins function}
\usage{
fragment_margins(configuration = NULL)
}
\arguments{
\item{configuration}{This is the 2D shape configuration}
}
\description{
fragment_margins function
}
\examples{
fragment_margins()
}
\keyword{fragment_margins}
