% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_efa.R
\name{i_efa}
\alias{i_efa}
\title{2D and 3D inverse EFA function}
\usage{
i_efa(efar, points = 200)
}
\arguments{
\item{efar}{Elliptical Fourier coefficients}

\item{points}{The number of points to approximate

Modified from the efourier_i function in the Mmocs package for R and Hayley Fancourt's 3D EFA code from Honours thesis}
}
\description{
2D and 3D inverse EFA function
}
\keyword{i_efa}
