% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthsort.R
\name{lengthsort}
\alias{lengthsort}
\title{A function to sort by bone length}
\usage{
lengthsort(sort, bone = "femur", side = "both", method = "Quartiles",
  measurements = NULL, sessiontempdir = NULL, stdout = FALSE,
  cutoff = 1.5)
}
\arguments{
\item{sort}{Sorted data for comparison}

\item{bone}{Specifies the bone type}

\item{side}{The bone type side ("left", "right", "both")}

\item{method}{The outlier method ("quartiles" or "standard deviations")}

\item{measurements}{The measurement types to be used}

\item{sessiontempdir}{Specifies temporary directory for analytical session if stdout is false}

\item{stdout}{If true, output will be data.frames only}

\item{cutoff}{The outlier cutoff value for either quartiles or standard deviations}
}
\description{
A function to sort by bone length
}
\examples{
lengthsort()
}
\keyword{lengthsort}
