% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.row.R
\name{m.row}
\alias{m.row}
\title{Match rows from different cols}
\usage{
m.row(xx, table)
}
\arguments{
\item{xx}{to be matched}

\item{table}{table to search}
}
\description{
Match rows from different cols
}
\examples{
m.row()
}
