% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.2d.invariant.R
\name{match.2d.invariant}
\alias{match.2d.invariant}
\title{two-dimensional pair-match function}
\usage{
match.2d.invariant(outlinedata = NULL, min = 1e+15, stdout = TRUE,
  sessiontempdir = NULL, oo = FALSE, iteration = 10,
  transformation = "rigid", cores = 1, test = "Segmented-Hausdorff",
  temporary_mean_specimen = 1, mean_iteration = 5, plot = FALSE,
  n_lowest_distances = 1, hide_distances = FALSE)
}
\arguments{
\item{outlinedata}{The outline data taken from outline.images()}

\item{min}{minimum distance for ICP}

\item{stdout}{If true, output will be data.frames only}

\item{sessiontempdir}{Specifies temporary directory for analytical session if stdout is false}

\item{iteration}{The number of iterations for Iterative Closest Point}

\item{transformation}{The type of Iterative Closest Point transformation ("Rigid", "Similarity", "Affine")}

\item{cores}{Number of cores for parallel processing}

\item{test}{Specifies the distance calculation ("Segmented-Hausdorff", "Hausdorff")}

\item{temporary_mean_specimen}{The specimen to be used as the temporary mean}

\item{mean_iteration}{The number of mean iterations}

\item{plot}{Plots results}

\item{n_lowest_distances}{The number of lowest distance matches to return as a potential match}

\item{hide_distances}{Hides the distance values in short lists to avoid analytical bias}
}
\description{
two-dimensional pair-match function
}
\examples{
match.2d.invariant()
}
\keyword{match.2d.invariant}
