% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.2d.R
\name{match.2d}
\alias{match.2d}
\title{match.3d Input Function
testing code}
\usage{
match.2d(spec1 = NULL, spec2 = NULL, plot = TRUE)
}
\arguments{
\item{spec1}{matrix}

\item{spec2}{matrix}

\item{plot}{TRUE}
}
\description{
match.3d Input Function
testing code
}
\examples{
match.2d()
}
\keyword{match.2d}

