% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsort.R
\name{metricsort}
\alias{metricsort}
\title{A function to sort by bone metrics}
\usage{
metricsort(sort, bone = "femur", side = "both", method = "Quartiles",
  measurements = NULL, sessiontempdir = NULL, cutoff = 1.5,
  output_options = c(TRUE, TRUE))
}
\arguments{
\item{sort}{Sorted data for comparison}

\item{bone}{Specifies the bone type}

\item{side}{The bone type side ("left", "right", "both")}

\item{method}{The outlier method ("quartiles" or "standard deviations")}

\item{measurements}{The measurement types to be used}

\item{sessiontempdir}{Specifies temporary directory for analytical session if stdout is false}

\item{cutoff}{The outlier cutoff value for either quartiles or standard deviations}

\item{output_options}{First index TRUE outputs to excel, second index TRUE outputs plot}
}
\description{
A function to sort by bone metrics
}
\examples{
metricsort()
}
\keyword{metricsort}
