% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OsteoSort.R
\name{OsteoSort}
\alias{OsteoSort}
\title{This function starts the OsteoSort app with shiny}
\usage{
OsteoSort()
}
\description{
This function starts the OsteoSort app with shiny
}
\examples{
OsteoSort()
}
\keyword{OsteoSort}

