% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outline.images.R
\name{outline.images}
\alias{outline.images}
\title{outline.images function}
\usage{
outline.images(imagelist1, imagelist2, threshold = 0.8, scale = TRUE,
  mirror = TRUE, npoints = 200, smooth_iterations = 1, nharmonics = 400)
}
\arguments{
\item{imagelist1}{List of image locations for first set}

\item{imagelist2}{List of image locations for second set}

\item{threshold}{Threshold value between white(1) and black(0)}

\item{scale}{Sets if a scale should be extracted from pixel dimensions}

\item{mirror}{If true mirrors imagelist2 to imagelist2}

\item{npoints}{The number of points in the inverse elliptical fourier analysis transformation}

\item{smooth_iterations}{The number of smoothing iterations in the elliptical fourier analysis}

\item{nharmonics}{The number of harmonics in elliptical fourier analysis}
}
\description{
outline.images function
}
\examples{
outline.images()
}
\keyword{outline.images}
