% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_align.R
\name{pca_align}
\alias{pca_align}
\title{Conducts PCA and returns scores}
\usage{
pca_align(configuration = NULL)
}
\arguments{
\item{first_configuration}{The first point cloud and the cloud to be mirroed}
}
\description{
Conducts PCA and returns scores
}
\examples{
pca_align()
}
\keyword{pca_align}
