% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.input.R
\name{pm.input}
\alias{pm.input}
\title{Pair-match Input Function}
\usage{
pm.input(bone = NULL, sort = NULL, measurement_standard = "standard",
  threshold = 1, measurements = NULL)
}
\arguments{
\item{bone}{Specifies the bone type}

\item{sort}{Data to be sorted}

\item{measurement_standard}{Specifies the measurement standards to use ("standard" or "supplemental")}

\item{threshold}{Threshold value for number of measurements per comparison}

\item{measurements}{The measurement types to be used}
}
\description{
Pair-match Input Function
}
\examples{
pm.input()
}
\keyword{pm.input}
