% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.ttest.R
\name{pm.ttest}
\alias{pm.ttest}
\title{Pair-match T.test Function}
\usage{
pm.ttest(ref = NULL, sort = NULL, sessiontempdir = NULL, stdout = TRUE,
  alphalevel = 0.1, power = TRUE, absolutevalue = TRUE,
  testagainstzero = FALSE, output_options = c(TRUE, FALSE), cores = 1,
  plot = FALSE)
}
\arguments{
\item{ref}{Reference data}

\item{sort}{Sorted data for comparison}

\item{sessiontempdir}{Specifies temporary directory for analytical session if stdout is false}

\item{stdout}{If true, output will be data.frames only}

\item{alphalevel}{Specifies alpha level}

\item{power}{If true, uses half-normal distribution power transformation}

\item{absolutevalue}{Uses absolute value for D-values if true}

\item{testagainstzero}{Uses 0 for mean if true}

\item{output_options}{Uses two true and/or false values to specify output file types. c(TRUE,) uses .txt file per specimen and C(,TRUE) uses .csv output files}

\item{cores}{Number of cores for parallel processing}

\item{plot}{Used internally for OsteoShiny, do not call directly!}
}
\description{
Pair-match T.test Function
}
\examples{
pm.ttest()
}
\keyword{pm.ttest}
