% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomstring.R
\name{randomstring}
\alias{randomstring}
\title{A Random String Function}
\usage{
randomstring(n = 1, length = 12)
}
\arguments{
\item{n}{The number of strings to return}

\item{length}{The length of the strings to return}
}
\description{
A Random String Function
}
\examples{
randomstring()
}
