% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtps.R
\name{readtps}
\alias{readtps}
\title{A function to import TPS data.}
\usage{
readtps(file)
}
\arguments{
\item{file}{File path for TPS data}
}
\description{
Modified from the opensource readland.tps function from the Geomorph package for R
}
\examples{
readtps()
}
\keyword{random}
