% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.multitest.R
\name{reg.multitest}
\alias{reg.multitest}
\title{reg.test Input Function
Function to produce combinations for associating elements with regression}
\usage{
reg.multitest(sort = NULL, ref = NULL, splitn = NULL,
  prediction_interval = 0.9, stdout = FALSE, sessiontempdir = NULL,
  output_options = c(TRUE, FALSE), cores = 1, test = TRUE,
  alphatest = TRUE, alphalevel = 0.05, plot = FALSE)
}
\arguments{
\item{sort}{Sorted data for comparison}

\item{ref}{Reference data}

\item{splitn}{The index location of each bone types measurements. Used internally from reg.input()}

\item{prediction_interval}{Specifies the prediction interval}

\item{stdout}{If true, output will be data.frames only}

\item{sessiontempdir}{Specifies temporary directory for analytical session if stdout is false}

\item{output_options}{Uses two true and/or false values to specify output file types. c(TRUE,) uses .txt file per specimen and C(,TRUE) uses .csv output files}

\item{cores}{Number of cores for parallel processing}

\item{test}{If true, PCA-CCA-Regression, if false, Simple Linear Regression}

\item{alphalevel}{Specifies alpha level}

\item{plot}{Used internally for OsteoShiny, do not call directly!}

\item{alphalevel}{Specifies alpha level}
}
\description{
reg.test Input Function
Function to produce combinations for associating elements with regression
}
\examples{
reg.multitest()
}
\keyword{reg.test}
