% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.test.R
\name{reg.test}
\alias{reg.test}
\title{reg.test Input Function
Function to produce combinations for associating elements with regression}
\usage{
reg.test(sort1 = NULL, sort2 = NULL, ref = NULL, predlevel = 0.9,
  stdout = FALSE, sessiontempdir = NULL)
}
\arguments{
\item{sort}{data to be sorted}
}
\description{
reg.test Input Function
Function to produce combinations for associating elements with regression
}
\examples{
reg.test()
}
\keyword{reg.test}

