% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_fragmented_margins.R
\name{remove_fragmented_margins}
\alias{remove_fragmented_margins}
\title{helper function to remove fragmented margins for Hausdorff}
\usage{
remove_fragmented_margins(first_configuration, second_configuration, indices)
}
\arguments{
\item{first_configuration}{first shape}

\item{second_configuration}{second shape}

\item{indices}{list of indices of fracture margins}
}
\description{
helper function to remove fragmented margins for Hausdorff
}
\examples{
remove_fragmented_margins()
}
\keyword{remove_fragmented_margins}
