% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmented_hausdorff_dist.R
\name{segmented_hausdorff_dist}
\alias{segmented_hausdorff_dist}
\title{segmented hausdorff distance function}
\usage{
segmented_hausdorff_dist(first_configuration, second_configuration,
  test = "segmented")
}
\arguments{
\item{first_configuration}{The first two-dimensional configuration}

\item{second_configuration}{The second two-dimensional configuration}

\item{test}{Specifies the distance calculation ("Segmented-Hausdorff", "Hausdorff")

Heavily modified from the opensource code in hausdorff_dist() function from the pracma package}
}
\description{
segmented hausdorff distance function
}
\examples{
segmented_hausdorff_dist()
}
\keyword{segmented_hausdorff_dist}
