% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftmatrices.R
\name{shiftmatrices}
\alias{shiftmatrices}
\title{Nearest neighbor matrices shifting function}
\usage{
shiftmatrices(first_configuration, second_configuration, threads = 1)
}
\arguments{
\item{first_configuration}{The first two-dimensional configuration}

\item{second_configuration}{The second two-dimensional configuration}

\item{threads}{The number of threads to use}
}
\description{
Nearest neighbor matrices shifting function
}
\examples{
shiftmatrices()
}
\keyword{shiftmatrices}
