% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writetps.R
\name{writetps}
\alias{writetps}
\title{A function to export TPS data.}
\usage{
writetps(shape_matrices, file)
}
\arguments{
\item{shape_matrices}{matrices of shape configurations}

\item{file}{File path for saving output}

\item{append}{TRUE FALSE create new file or append existing}
}
\description{
Modified from the opensource writeland.tps function from the Geomorph package for R
}
\examples{
writetps()
}
\keyword{random}
