% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antestat.regtest.R
\name{antestat.regtest}
\alias{antestat.regtest}
\title{antestat.regtest Input Function}
\usage{
antestat.regtest(sort = NULL, ref = NULL, sessiontempdir = NULL,
  output_options = c(TRUE, FALSE), prediction_interval = 0.95, tails = 2,
  alphalevel = 0.05, alphatest = TRUE, threads = 1)
}
\arguments{
\item{sort}{The antemortem stature and postmorterm measurement for comparison}

\item{ref}{The reference sample used in the regression model}

\item{sessiontempdir}{Specifies temporary directory for analytical session}

\item{output_options}{C(TRUE,FALSE) First logic specifies excel output, second specifies plot output}

\item{prediction_interval}{The prediction interval level for exclusion}

\item{tails}{The number of tails for the t-distribution}

\item{alphalevel}{The alpha level for exclusion}

\item{alphatest}{If TRUE specifies the use of alphalevels for exclusion. If FALSE specifies the use of prediction intervals}

\item{threads}{The number of threads to use}
}
\description{
Function to produce a p-value evaluating the strength of evidence for comparing an antemortem stature to postmortem measurement
}
\examples{
antestat.regtest()
}
\keyword{antestat}
