% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art.input.R
\name{art.input}
\alias{art.input}
\title{Articulation-match Input Function}
\usage{
art.input(bone = NULL, sort = NULL, template = "single")
}
\arguments{
\item{bone}{bone}

\item{sort}{sort}

\item{template}{template}

\item{reference}{reference}
}
\description{
This function takes input from the user to be articulation-matched and generates
a reference database for each possible combination based on the
number of measurements available. It's designed to work as a single user input or
with the multiple comparison standardized csv format. It accounts for NA values automatically.
}
\examples{
art.input()
}
\keyword{art.input}

