% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art.ttest.R
\name{art.ttest}
\alias{art.ttest}
\title{Articulation comparison t-test Function}
\usage{
art.ttest(ref = NULL, sort = NULL, sessiontempdir = NULL, cores = 1,
  alphalevel = 0.1, absolutevalue = TRUE, testagainstzero = FALSE,
  output_options = c(TRUE, FALSE), power = TRUE)
}
\arguments{
\item{ref}{Reference data}

\item{sort}{Sorted data for comparison}

\item{sessiontempdir}{Specifies temporary directory for analytical session}

\item{cores}{Number of cores for parallel processing}

\item{alphalevel}{Specifies alpha level}

\item{absolutevalue}{Uses absolute value for D-values if true}

\item{testagainstzero}{Uses 0 for mean if true}

\item{output_options}{C(TRUE,FALSE) First logic specifies excel output, second specifies plot output}

\item{power}{If true, uses half-normal distribution power transformation}
}
\description{
Articulation comparison t-test Function
}
\examples{
art.ttest()
}
\keyword{art.ttest}
