% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art.ttest.R
\name{art.ttest}
\alias{art.ttest}
\title{Articulation comparison t-test Function}
\usage{
art.ttest(ref = NULL, sort = NULL, sessiontempdir = NULL, threads = 1,
  alphalevel = 0.1, absolutevalue = TRUE, testagainstzero = FALSE,
  output_options = c(TRUE, FALSE), power = TRUE, tails = 2)
}
\arguments{
\item{ref}{Reference data}

\item{sort}{Sorted data for comparison}

\item{sessiontempdir}{Specifies temporary directory for analytical session}

\item{threads}{The number of threads to use}

\item{alphalevel}{The alpha level for exclusion}

\item{absolutevalue}{if TRUE uses absolute value for D-values}

\item{testagainstzero}{if TRUE uses 0 for sample mean}

\item{output_options}{C(TRUE,FALSE) First logic specifies excel output, second specifies plot output}

\item{power}{If TRUE uses half-normal distribution power transformation}

\item{tails}{The number of tails for the t-distribution}
}
\description{
Articulation comparison t-test Function
}
\examples{
art.ttest()
}
\keyword{art.ttest}
