% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/art.ttest.R
\name{art.ttest}
\alias{art.ttest}
\title{Articulation-match T.test Function}
\usage{
art.ttest(refdata = NULL, sortdata = NULL, template = "single",
  sessiontempdir = NULL, stdout = TRUE, alphalevel = 0.1,
  absolutevalue = TRUE, a = FALSE)
}
\arguments{
\item{refdata}{refdata}

\item{sortdata}{sortdata}

\item{template}{template}

\item{alltest}{alltest}

\item{sessiontemp}{sessiontemp}
}
\description{
This function takes the combinations and reference data generated from the art.input() function 
and runs a t.test for each combination. If the results are not statistically significant
the combination is considered a possible match and saved in two formats, a single .txt file with all matches
and a .csv file with matches. If it's not a match the combination is saved to a .csv file with non-matches
}
\examples{
art.ttest()
}
\keyword{art.ttest}

