% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hausdorff_dist.R
\name{hausdorff_dist}
\alias{hausdorff_dist}
\title{hausdorff distance function}
\usage{
hausdorff_dist(first_configuration, second_configuration,
  test = "Segmented-Hausdorff", n_regions = 0, dist = "maximum",
  cores = 1, indices = NULL)
}
\arguments{
\item{first_configuration}{The first two-dimensional configuration}

\item{second_configuration}{The second two-dimensional configuration}

\item{test}{Specifies the distance calculation ("Segmented-Hausdorff", "Hausdorff", "Uni-Hausdorff")}

\item{n_regions}{Specifies the number of regions for Segmented-Hausdorff}

\item{dist}{Specifies distance per region, either maximum or average distance}

\item{cores}{The number of threads/cores to use with RcppParallel

Heavily modified from the opensource code in hausdorff_dist() function from the pracma package}
}
\description{
hausdorff distance function
}
\examples{
hausdorff_dist()
}
\keyword{segmented_hausdorff_dist}
