% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthsort.R
\name{lengthsort}
\alias{lengthsort}
\title{A function to sort by bone length}
\usage{
lengthsort(file, bone = "femur", side = "both", outfile = NULL,
  method = "sd")
}
\arguments{
\item{file}{test}

\item{population}{test}
}
\description{
A function to sort by bone length
}
\examples{
lengthsort()
}
\keyword{random}

