% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.3d.R
\name{match.3d}
\alias{match.3d}
\title{match.3d Input Function
testing code}
\usage{
match.3d(spec1 = NULL, spec2 = NULL, min = 10, iter = 1000,
  trans = "rigid", plot = TRUE, aspect = c(4, 1, 1), GPA = TRUE,
  meanshape = FALSE, regression = FALSE)
}
\arguments{
\item{spec1}{First specimen configuration}

\item{spec2}{Second specimen configuration}

\item{min}{minimum distance between points in ICP}

\item{iter}{number of iterations for ICP}

\item{trans}{type of transformation (affine, rigid, similarity)}

\item{plot}{produce plot}

\item{aspect}{aspect of plot}

\item{GPA}{use GPA or just ICP}

\item{meanshape}{calculate mean shape}
}
\description{
match.3d Input Function
testing code
}
\examples{
match.3d()
}
\keyword{match.3d}

