% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_function.R
\name{output_function}
\alias{output_function}
\title{output_excel function}
\usage{
output_function(hera1)
}
\arguments{
\item{results}{Only used to call internally}
}
\description{
output_excel function
}
