% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.input.R
\name{pm.input}
\alias{pm.input}
\title{Pair-match Input Function}
\usage{
pm.input(bone = NULL, sort = NULL, template = "single_standard",
  tresh = 2, measurements = NULL)
}
\arguments{
\item{bone}{bone name}

\item{sort}{data to be sorted}

\item{template}{single of multiple}

\item{tresh}{treshhold value}

\item{reference}{reference data}
}
\description{
This function takes input from the user to be pair-matched and generates
a reference database for each possible combination based on the
number of measurements available. It's designed to work as a single user input or
with the multiple comparison standardized csv format. It accounts for NA values automatically.
}
\examples{
pm.input()
}
\keyword{pm.input}

