% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.ttest.R
\name{pm.ttest}
\alias{pm.ttest}
\title{Pair-match T.test Function}
\usage{
pm.ttest(ref = NULL, sort = NULL, sessiontempdir = NULL,
  alphalevel = 0.1, power = TRUE, absolutevalue = TRUE,
  testagainstzero = FALSE, output_options = c(TRUE, FALSE), threads = 1,
  tails = 1)
}
\arguments{
\item{ref}{Reference data}

\item{sort}{Sorted data for comparison}

\item{sessiontempdir}{Specifies temporary directory for analytical session}

\item{alphalevel}{The alpha level for exclusion}

\item{power}{If TRUE uses half-normal distribution power transformation}

\item{absolutevalue}{if TRUE uses absolute value for D-values}

\item{testagainstzero}{if TRUE uses 0 for sample mean}

\item{output_options}{C(TRUE,FALSE) First logic specifies excel output, second specifies plot output}

\item{threads}{The number of threads to use}

\item{tails}{The number of tails for the t-distribution}
}
\description{
Pair-match T.test Function
}
\examples{
pm.ttest()
}
\keyword{pm.ttest}
