% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm.ttest.R
\name{pm.ttest}
\alias{pm.ttest}
\title{Pair-match T.test Function}
\usage{
pm.ttest(refdata = NULL, sortdata = NULL, template = "single",
  sessiontempdir = NULL, stdout = TRUE, alphalevel = 0.1,
  absolutevalue = TRUE, a = FALSE)
}
\arguments{
\item{refdata}{refdata}

\item{sortdata}{sortdata}

\item{tempplate}{template}

\item{alltest}{alltest}

\item{sessiontemp}{sessiontemp}
}
\description{
This function takes the combinations and reference data generated from the pm.input() function 
and runs a t.test for each combination. If the results are not statistically significant
the combination is considered a possible match and saved in two formats, a single .txt file with all matches
and a .csv file with matches. If it's not a match the combination is saved to a .csv file with non-matches
}
\examples{
pm.ttest()
}
\keyword{pm.ttest}

