% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomstring.R
\name{randomstring}
\alias{randomstring}
\title{A Random String Function}
\usage{
randomstring(n = 1, length = 12)
}
\arguments{
\item{n}{how many}

\item{length}{length of string}
}
\description{
This function allows you to generate a random string of length 12 for 
generating temporary directories and files. It's used to create the temporary local 
environments for saving and zipping results
}
\examples{
randomstring()
}
\keyword{random}

