% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.input.R
\name{reg.input}
\alias{reg.input}
\title{reg.match Input Function
Function to produce combinations for associating elements with regression}
\usage{
reg.input(sort1 = NULL, sort2 = NULL, template = "single_standard")
}
\arguments{
\item{sort}{data to be sorted}
}
\description{
reg.match Input Function
Function to produce combinations for associating elements with regression
}
\examples{
reg.input(sort1 = c("ID1","Left","femur",23,24,25,NA,NA,NA,NA,NA), sort2 = c("ID2","Left","tibia",23,24,25,NA))
reg.input()
}
\keyword{reg.input}

