% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.multitest.R
\name{reg.multitest}
\alias{reg.multitest}
\title{reg.test Input Function
Function to produce combinations for associating elements with regression}
\usage{
reg.multitest(sort = NULL, ref = NULL, splitn = NULL,
  prediction_interval = 0.9, sessiontempdir = NULL,
  output_options = c(TRUE, FALSE), threads = 1, test = TRUE,
  alphatest = TRUE, alphalevel = 0.05)
}
\arguments{
\item{sort}{Sorted data for comparison}

\item{ref}{Reference data}

\item{splitn}{The index location of each bone types measurements. Used internally from reg.input()}

\item{prediction_interval}{Specifies the prediction interval}

\item{sessiontempdir}{Specifies temporary directory for analytical session}

\item{output_options}{C(TRUE,FALSE) First logic specifies excel output, second specifies plot output}

\item{threads}{The number of threads to use}

\item{test}{If true, PCA-CCA-Regression, if false, Simple Linear Regression}

\item{alphalevel}{The alpha level for exclusion}
}
\description{
reg.test Input Function
Function to produce combinations for associating elements with regression
}
\examples{
reg.multitest()
}
\keyword{reg.test}
