% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsort.R
\name{statsort}
\alias{statsort}
\title{A function estimate and sort by stature}
\usage{
statsort(file, bone = "femur", side = "both", population = "trotter-any",
  outfile = NULL, method = "sd")
}
\arguments{
\item{file}{test}

\item{population}{teast}
}
\description{
Equation reference from the Forensic Data Bank
}
\examples{
statsort()
}
\keyword{random}

