% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsort.R
\name{statsort}
\alias{statsort}
\title{A function estimate and sort by stature}
\usage{
statsort(sort, bone = "femur", side = "both",
  population = "trotter-any-male", method = "Quartiles",
  measurements = NULL, sessiontempdir = NULL, cutoff = 1.5,
  output_options = c(TRUE, TRUE), metric = "mm")
}
\arguments{
\item{sort}{Sorted data for comparison}

\item{bone}{Specifies the bone type ("left", "right", "both")}

\item{side}{The bone type side}

\item{population}{The reference sample for stature estimation}

\item{method}{The outlier method ("quartiles" or "standard deviations")}

\item{measurements}{The measurement types to be used}

\item{sessiontempdir}{Specifies temporary directory for analytical session}

\item{cutoff}{The outlier cutoff value for either quartiles or standard deviations}

\item{output_options}{First index TRUE outputs to excel, second index TRUE outputs plot}

\item{metric}{Specifies millimeters (mm), centimeters (cm), or inches (in) for stature}
}
\description{
Equation reference from the Forensic Data Bank
}
\examples{
statsort()
}
\keyword{random}
