% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writetps.R
\name{writetps}
\alias{writetps}
\title{A function to export TPS data.}
\usage{
writetps(A, file, scale = NULL)
}
\arguments{
\item{file}{test}
}
\description{
Modified from the opensource writeland.tps function from the Geomorph package for R
}
\examples{
writetps()
}
\keyword{random}

